# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AIOManager integration API",
    "version": "1.0.0",
    "description": "Application 2.0.0. All timestamps are Unix milliseconds; membership also includes local wall time and IANA timezone. HTTPS is required for remote use. No billing or IPTV operations. Persist each mutation intent/key before sending. Poll jobs until verified; failed is not proof of no external effect."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "AIOManager API key"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Status": {
        "type": "object",
        "properties": {
          "capabilities": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          },
          "writePaused": {
            "type": "boolean"
          },
          "ownerWritePaused": {
            "type": "boolean"
          },
          "writerReady": {
            "type": "boolean"
          },
          "lastScanAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "lastBackupAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "safeMode": {
            "type": "boolean"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "accounts": {
            "type": "object",
            "properties": {
              "staged": {
                "type": "integer"
              },
              "active": {
                "type": "integer"
              },
              "offboarding": {
                "type": "integer"
              }
            },
            "required": [
              "staged",
              "active",
              "offboarding"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "capabilities",
          "writePaused",
          "ownerWritePaused",
          "writerReady",
          "lastScanAt",
          "lastBackupAt",
          "safeMode",
          "version",
          "accounts"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/status": {
      "get": {
        "operationId": "get__status",
        "summary": "Sync and application backup status",
        "parameters": [],
        "x-required-scopes": [
          "read"
        ],
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Status"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/accounts": {
      "get": {
        "operationId": "get__accounts",
        "summary": "List managed accounts without passwords or addon URLs",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "view",
            "in": "query",
            "schema": {
              "enum": [
                "all",
                "expired"
              ],
              "default": "all"
            }
          }
        ],
        "x-required-scopes": [
          "read"
        ],
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accounts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Account"
                      }
                    },
                    "nextCursor": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "uuid"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "accounts",
                    "nextCursor"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post__accounts",
        "summary": "Stage an existing Stremio login; does not register a Stremio account or activate sync",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            },
            "description": "Durable receipt action: create. Same key and intent replays; different content returns 409."
          }
        ],
        "x-required-scopes": [
          "read",
          "accounts:write"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccount"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedAccount"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "operationId": "get__accounts__id_",
        "summary": "Read account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-required-scopes": [
          "read"
        ],
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{id}/execution": {
      "get": {
        "operationId": "get__accounts__id__execution",
        "summary": "Read current account and latest job, or a removal tombstone",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-required-scopes": [
          "read"
        ],
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Execution"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts/{id}/addons": {
      "get": {
        "operationId": "get__accounts__id__addons",
        "summary": "Read saved addon configuration; URLs may contain secrets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-required-scopes": [
          "read",
          "configuration:read"
        ],
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post__accounts__id__addons",
        "summary": "Save individual addon configuration",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            },
            "description": "Durable receipt action: addons. Same key and intent replays; different content returns 409."
          }
        ],
        "x-required-scopes": [
          "read",
          "accounts:write",
          "configuration:read"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountAddons"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts/{id}/name": {
      "post": {
        "operationId": "post__accounts__id__name",
        "summary": "Rename account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            },
            "description": "Durable receipt action: name. Same key and intent replays; different content returns 409."
          }
        ],
        "x-required-scopes": [
          "read",
          "accounts:write"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "expectedVersion": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  }
                },
                "required": [
                  "expectedVersion",
                  "name"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResult"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{id}/membership": {
      "post": {
        "operationId": "post__accounts__id__membership",
        "summary": "Set dated or lifetime membership; a past cutoff expires the account",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            },
            "description": "Durable receipt action: membership. Same key and intent replays; different content returns 409."
          }
        ],
        "x-required-scopes": [
          "read",
          "accounts:write"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Membership"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResult"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/bulk-membership": {
      "post": {
        "operationId": "post__accounts_bulk_membership",
        "summary": "Set membership for a versioned account selection",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            },
            "description": "Durable receipt action: bulk-membership. Same key and intent replays; different content returns 409."
          }
        ],
        "x-required-scopes": [
          "read",
          "accounts:write"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkMembership"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Saved or read successfully. A non-null jobId means provider work must be polled; this is not proof of sync completion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResult"
                }
              }
            }
          },
          "default": {
            "description": "Sanitized error; 401 invalid/expired/revoked key, 403 missing scope, 409 conflicting state/request, 429 rate limit. An interrupted or 5xx mutation has an uncertain outcome: recover its receipt or retry the same key.",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}